delete the right Binder.ts file
